# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_the-monkeys_freerangenotify_internal_domain_user.CategoryPreference:
    properties:
      enabled:
        type: boolean
      enabled_channels:
        items:
          type: string
        type: array
    type: object
...
  github_com_the-monkeys_freerangenotify_internal_domain_user.Preferences:
    properties:
      categories:
        additionalProperties:
          $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_domain_user.CategoryPreference'
        description: Category-specific preferences
        type: object
      daily_limit:
        description: Max notifications per day
        type: integer
      discord_enabled:
        description: Phase 3
        type: boolean
      dnd:
        description: Global Do Not Disturb
        type: boolean
      email_enabled:
        type: boolean
      push_enabled:
        type: boolean
      quiet_hours:
        $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_domain_user.QuietHours'
      slack_enabled:
        description: Phase 3
        type: boolean
      sms_enabled:
        type: boolean
      throttle:
        additionalProperties:
          $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_domain_user.ThrottleConfig'
        description: Per-channel throttle overrides (Phase 2)
        type: object
      whatsapp_enabled:
        description: Phase 3
        type: boolean
    type: object
  github_com_the-monkeys_freerangenotify_internal_domain_user.QuietHours:
    properties:
      enabled:
        type: boolean
      end:
        description: 'Format: "HH:MM"'
        type: string
      start:
        description: 'Format: "HH:MM"'
        type: string
    type: object
  github_com_the-monkeys_freerangenotify_internal_domain_user.ThrottleConfig:
    properties:
      max_per_day:
        type: integer
      max_per_hour:
        type: integer
    type: object
...
  github_com_the-monkeys_freerangenotify_internal_interfaces_http_dto.AddDeviceRequest:
    properties:
      platform:
        enum:
        - ios
        - android
        - web
        type: string
      token:
        type: string
    required:
    - platform
    - token
    type: object
...
  github_com_the-monkeys_freerangenotify_internal_interfaces_http_dto.BulkCreateUserRequest:
    properties:
      users:
        items:
          $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_interfaces_http_dto.CreateUserRequest'
        maxItems: 1000
        minItems: 1
        type: array
    required:
    - users
    type: object
  github_com_the-monkeys_freerangenotify_internal_interfaces_http_dto.BulkCreateUserResponse:
    properties:
      created:
        type: integer
      errors:
        items:
          $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_interfaces_http_dto.BulkUserError'
        type: array
      total:
        type: integer
    type: object
...
  github_com_the-monkeys_freerangenotify_internal_interfaces_http_dto.BulkUserError:
    properties:
      email:
        type: string
      index:
        type: integer
      message:
        type: string
    type: object
...
  github_com_the-monkeys_freerangenotify_internal_interfaces_http_dto.CreateUserRequest:
    properties:
      email:
        type: string
      external_id:
        type: string
      language:
        type: string
      phone:
        type: string
      preferences:
        $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_domain_user.Preferences'
      timezone:
        type: string
      user_id:
        type: string
      webhook_url:
        type: string
    type: object
...
  github_com_the-monkeys_freerangenotify_internal_interfaces_http_dto.UpdatePreferencesRequest:
    properties:
      categories:
        additionalProperties:
          $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_domain_user.CategoryPreference'
        type: object
      daily_limit:
        type: integer
      discord_enabled:
        description: Phase 5
        type: boolean
      dnd:
        type: boolean
      email_enabled:
        type: boolean
      push_enabled:
        type: boolean
      quiet_hours:
        $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_domain_user.QuietHours'
      slack_enabled:
        description: Phase 5
        type: boolean
      sms_enabled:
        type: boolean
      whatsapp_enabled:
        description: Phase 5
        type: boolean
    type: object
...
  github_com_the-monkeys_freerangenotify_internal_interfaces_http_dto.UpdateUserRequest:
    properties:
      email:
        type: string
      external_id:
        type: string
      language:
        type: string
      phone:
        type: string
      preferences:
        $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_domain_user.Preferences'
      timezone:
        type: string
      webhook_url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@freerangenotify.com
    name: API Support
  description: High-performance notification service with multi-channel delivery support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FreeRangeNotify API
  version: 0.1.9-alpha
paths:
...
  /v1/users:
    get:
      description: List users for the authenticated application with pagination and
        filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by email
        in: query
        name: email
        type: string
      - description: Filter by timezone
        in: query
        name: timezone
        type: string
      - description: Filter by language
        in: query
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new subscriber/user within an application
      parameters:
      - description: User creation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_interfaces_http_dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new user
      tags:
      - Users
  /v1/users/{id}:
    delete:
      description: Permanently remove a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a user
      tags:
      - Users
    get:
      description: Retrieve a single user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get a user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update an existing user's information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_interfaces_http_dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a user
      tags:
      - Users
  /v1/users/{id}/devices:
    get:
      description: Retrieve all registered devices for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user devices
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Register a push notification device token for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Device registration request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_interfaces_http_dto.AddDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add a device to a user
      tags:
      - Users
  /v1/users/{id}/devices/{device_id}:
    delete:
      description: Unregister a push notification device token
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Remove a device from a user
      tags:
      - Users
  /v1/users/{id}/preferences:
    get:
      description: Retrieve notification preferences for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user notification preferences
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update channel preferences, quiet hours, and DND settings for a
        user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Preferences update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_interfaces_http_dto.UpdatePreferencesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update user notification preferences
      tags:
      - Users
  /v1/users/{id}/subscriber-hash:
    get:
      description: Generate an HMAC subscriber hash for SSE authentication
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get subscriber hash
      tags:
      - Users
  /v1/users/bulk:
    post:
      consumes:
      - application/json
      description: Create multiple users in a single request
      parameters:
      - description: Bulk user creation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_interfaces_http_dto.BulkCreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_interfaces_http_dto.BulkCreateUserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Bulk create users
      tags:
      - Users
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Application API key for tenant-facing endpoints
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and your API key
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"